// src/interfaces.d.ts
export interface TaxonomyNode {
	key: string;
	value: string;
	path: string;
	children?: TaxonomyNode[];
	hasChildren: boolean;
}

export interface EntryGetOptions {
	id: string;
	language?: string;
	linkDepth?: number;
}

export interface EntryListOptions {
	contentTypeId?: string;
	language?: string;
	pageOptions?: PageOptions;
	order?: string[];
	linkDepth?: number;
	fields?: string[];
}

interface TaxonomyGetOptions {
	order?: 'alphabetical' | 'defined';
	childDepth?: number;
	language?: string;
}

export interface TaxonomyGetNodeByKeyOptions extends TaxonomyGetOptions {
	key: string;
}

export interface TaxonomyGetNodeByPathOptions extends TaxonomyGetOptions {
	path: string;
}

export interface TaxonomyResolveChildrenOptions extends TaxonomyGetOptions {
	node: string | TaxonomyNode;
}
// ...
export interface ITaxonomyOperations {
	getNodeByKey(key: string | TaxonomyGetNodeByKeyOptions): Promise<TaxonomyNode>;
	getNodeByPath(path: string | TaxonomyGetNodeByPathOptions): Promise<TaxonomyNode>;
	resolveChildren(node: string | TaxonomyNode | TaxonomyResolveChildrenOptions): Promise<TaxonomyNode>;
}

export interface IHttpClient {
	request<T>(url: string, request?: RequestInit): Promise<T>;
}

export interface IParamsProvider {
	getParams(): ClientParams;
}

export interface ContensisClient extends IParamsProvider {
	entries: IEntryOperations;
	contentTypes: IContentTypeOperations;
	project: IProjectOperations;
	taxonomy: ITaxonomyOperations;
}

export type VersionStatus = 'published' | 'latest';

export interface Config {
	rootUrl?: string;
	accessToken?: string;
	projectId?: string;
	language?: string;
	versionStatus?: VersionStatus;
	pageSize?: number;
}

export interface ClientParams {
	rootUrl: string;
	accessToken: string;
	language: string;
	versionStatus: VersionStatus;
	projectId: string;
	pageIndex: number;
	pageSize: number;
}